Added Reaction edit from.

+++ src/main/resources/static/js/jquery.boot.js
@@ -28,8 +28,20 @@ $(document).ready(function () {
         }
     });
 
+    $("#panels").on("change", "#update-reaction-image",function () {
+        let file = $("#update-reaction-image").get(0).files[0];
+
+        if (file) {
+            let reader = new FileReader();
+
+            reader.onload = function () {
+                $("#update-reaction-image-preview").attr("src", reader.result);
+            }
+            reader.readAsDataURL(file);
+        }
+    });
+
     $("#posts").on("click", "button[name='post-delete']", function () {
-        console.log(1);
         let id = $(this).val();
         if (confirm('Post ' + id + ' will be removed.')) {
             deletePost(id);
@@ -46,8 +58,17 @@ $(document).ready(function () {
 
     $("#reactions").on("click", "button[name='reaction-delete']", function () {
         let id = $(this).val();
-        deleteReaction(id);
+        if (confirm('Are You sure You want to delete this reaction?')) {
+            deleteReaction(id);
+        }
     });
+
+    $("#reactions").on("click", "button[name='reaction-edit']", function () {
+        $("#reactions-create").remove();
+
+        let id = $(this).val();
+        loadReactionDetails(id);
+    })
 });
 
 function loadPosts() {
@@ -86,6 +107,17 @@ function loadReactionPanelData() {
                appendReactionDataToPanel(reaction);
            });
        }
+    });
+}
+
+function loadReactionDetails(id) {
+    $.ajax({
+        type: "GET",
+        url: "/api/reactions/" + id,
+        dataType: "JSON",
+        success: function (reaction) {
+            appendReactionEditFormToPanel(reaction);
+        }
     });
 }
 
@@ -158,4 +190,33 @@ function appendReactionDataToPanel(reaction) {
             '</td>' +
         '</tr>'
     )
+}
+
+function appendReactionEditFormToPanel(reaction) {
+    let editForm = $("#panels #reaction-edit");
+    editForm.remove();
+
+    $("#panels").append(
+        '<div id="reaction-edit" class="max-w-md mx-auto rounded-xl mb-4 bg-white p-4 grid grid-flow-row auto-rows-max shadow">' +
+            '<div class="border p-4 rounded-lg">' +
+                '<span class="block mb-2 text-sm uppercase font-medium text-gray-900 dark:text-white text-center">' +
+                    'Edit reaction: ' + reaction.name +
+                '</span>' +
+                '<form action="/api/reactions" method="PUT" encType="multipart/form-data" id="update-reaction-form">' +
+                    '<div class="mb-6">' +
+                        '<label for="update-reaction-name" class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Name</label>' +
+                        '<input type="text" value="' + reaction.name + '" id="update-reaction-name" name="name" class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500" required>' +
+                    '</div>' +
+                    '<div class="mb-6">' +
+                        '<img id="update-reaction-image-preview" src="">' +
+                        '<label for="update-reaction-image" class="block mb-2 text-sm font-medium text-gray-900 dark:text-white">Image</label>' +
+                        '<input type="file" id="update-reaction-image" name="image" class="bg-gray-50 border border-gray-300 text-gray-900 text-sm rounded-lg focus:ring-blue-500 focus:border-blue-500 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500" required>' +
+                    '</div>' +
+                    '<button type="submit" class="text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm w-full sm:w-auto px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">' +
+                        'Edit' +
+                    '</button>' +
+                '</form>' +
+            '</div>' +
+        '</div>'
+    );
 }
